+++ statistics/handle/handletest/BUILD.bazel
@@ -9,7 +9,7 @@ go_test(
     ],
     flaky = True,
     race = "on",
-    shard_count = 38,
+    shard_count = 37,
     deps = [
         "//config",
         "//domain",
