+++ statistics/handle/handletest/statstest/BUILD.bazel
@@ -9,16 +9,13 @@ go_test(
     ],
     flaky = True,
     race = "on",
-    shard_count = 9,
+    shard_count = 8,
     deps = [
         "//config",
         "//parser/model",
-        "//planner/cardinality",
         "//statistics/handle/internal",
         "//testkit",
         "//testkit/testsetup",
-        "//types",
-        "@com_github_pingcap_failpoint//:failpoint",
         "@com_github_stretchr_testify//require",
         "@org_uber_go_goleak//:goleak",
     ],
